Movie/requirements.txt
appdirs==1.4.4
babelfish==0.5.5
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==8.0.1
colorama==0.4.4
colorful==0.5.4
cryptography==3.3.1
decorator==5.0.9
defusedxml==0.6.0
Django==2.2
django-allauth==0.44.0
django-crispy-forms==1.10.0
django-crontab==0.7.1
django-filter==2.4.0
django-tqdm==1.0.0
dogpile.cache==1.1.3
enzyme==0.4.1
feather-format==0.4.1
future==0.18.2
guessit==3.3.1
halo==0.0.31
idna==2.10
log-symbols==0.0.14
mysqlclient==2.0.3
numpy==1.20.3
oauthlib==3.1.0
pandas==1.1.5
pbr==5.6.0
Pillow==8.0.1
PTable==0.9.2
pyarrow==4.0.1
pycparser==2.20
PyJWT==1.7.1
pysrt==1.1.2
python-dateutil==2.8.1
python-environ==0.4.54
python3-openid==3.2.0
pytz==2020.4
rarfile==4.0
rebulk==3.0.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
spinners==0.0.24
sqlparse==0.4.1
stevedore==3.3.0
subliminal==2.1.0
sweetify==2.0.0
termcolor==1.1.0
tmdbv3api==1.7.5
tqdm==4.54.1
urllib3==1.26.2
